+++ packages/react-form-state/src/components/Nested.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import get from 'lodash/get';
 import {memoize, bind} from 'lodash-decorators';
 
-import {FieldDescriptor, FieldDescriptors} from '../types';
+import {FieldDescriptor, FieldDescriptors, ValueMapper} from '../types';
 import {mapObject} from '../utilities';
 
 interface Props<Fields> {
@@ -53,17 +53,20 @@ export default class Nested<Fields> extends React.Component<
   @memoize()
   @bind()
   private handleChange<Key extends keyof Fields>(key: Key) {
-    return (newValue: Fields[Key]) => {
+    return (newValue: Fields[Key] | ValueMapper<Fields>) => {
       const {
-        field: {value: existingItem, onChange},
+        field: {onChange},
       } = this.props;
 
-      const newItem = {
-        ...(existingItem as any),
-        [key]: newValue,
-      };
-
-      onChange(newItem);
+      onChange(value => {
+        return {
+          ...(value as any),
+          [key]:
+            typeof newValue === 'function'
+              ? newValue(value[key as string])
+              : newValue,
+        };
+      });
     };
   }
 }