+++ packages/react-form-state/src/components/tests/Nested.test.tsx
@@ -131,6 +131,143 @@ describe('<Nested />', () => {
     expect(renderPropArgs.department.error).toBe(field.error.department);
   });
 
+  it('Does not have race condition with multiple onChange calls', () => {
+    const product = {
+      title: faker.commerce.productName(),
+      department: faker.commerce.department(),
+    };
+
+    const renderSpy = jest.fn(({title, department}) => {
+      return (
+        <>
+          <Input label="title" {...title} />
+          <Input label="department" {...department} />
+        </>
+      );
+    });
+
+    mount(
+      <FormState initialValues={{product}}>
+        {({fields}) => {
+          return (
+            <FormState.Nested field={fields.product}>
+              {renderSpy}
+            </FormState.Nested>
+          );
+        }}
+      </FormState>,
+    );
+
+    const {title, department} = lastCallArgs(renderSpy);
+
+    const newTitle = faker.commerce.productName();
+    const newDepartment = faker.commerce.department();
+
+    title.onChange(newTitle);
+    department.onChange(newDepartment);
+
+    const updatedFields = lastCallArgs(renderSpy);
+
+    expect(updatedFields.title.value).toBe(newTitle);
+    expect(updatedFields.department.value).toBe(newDepartment);
+  });
+
+  it('Does not have race condition when using Nested -> Nested', () => {
+    const product = {
+      nested: {
+        title: faker.commerce.productName(),
+        department: faker.commerce.department(),
+      },
+    };
+
+    const renderSpy = jest.fn(({title, department}) => {
+      return (
+        <>
+          <Input label="title" {...title} />
+          <Input label="department" {...department} />
+        </>
+      );
+    });
+
+    mount(
+      <FormState initialValues={{product}}>
+        {({fields}) => {
+          return (
+            <FormState.Nested field={fields.product}>
+              {nestedFields => (
+                <FormState.Nested field={nestedFields.nested}>
+                  {renderSpy}
+                </FormState.Nested>
+              )}
+            </FormState.Nested>
+          );
+        }}
+      </FormState>,
+    );
+
+    const {title, department} = lastCallArgs(renderSpy);
+
+    const newTitle = faker.commerce.productName();
+    const newDepartment = faker.commerce.department();
+
+    title.onChange(newTitle);
+    department.onChange(newDepartment);
+
+    const updatedFields = lastCallArgs(renderSpy);
+
+    expect(updatedFields.title.value).toBe(newTitle);
+    expect(updatedFields.department.value).toBe(newDepartment);
+  });
+
+  it('Does not have race condition when using List -> Nested', () => {
+    const product = [
+      {
+        nested: {
+          title: faker.commerce.productName(),
+          department: faker.commerce.department(),
+        },
+      },
+    ];
+
+    const renderSpy = jest.fn(({title, department}) => {
+      return (
+        <>
+          <Input label="title" {...title} />
+          <Input label="department" {...department} />
+        </>
+      );
+    });
+
+    mount(
+      <FormState initialValues={{product}}>
+        {({fields}) => {
+          return (
+            <FormState.List field={fields.product}>
+              {nestedFields => (
+                <FormState.Nested field={nestedFields.nested}>
+                  {renderSpy}
+                </FormState.Nested>
+              )}
+            </FormState.List>
+          );
+        }}
+      </FormState>,
+    );
+
+    const {title, department} = lastCallArgs(renderSpy);
+
+    const newTitle = faker.commerce.productName();
+    const newDepartment = faker.commerce.department();
+
+    title.onChange(newTitle);
+    department.onChange(newDepartment);
+
+    const updatedFields = lastCallArgs(renderSpy);
+
+    expect(updatedFields.title.value).toBe(newTitle);
+    expect(updatedFields.department.value).toBe(newDepartment);
+  });
+
   it('Does not re-render when children have not changed', () => {
     const titleSpy = jest.fn(() => null);
     const adjectiveSpy = jest.fn(({adjective}) => <Input {...adjective} />);
@@ -158,9 +295,8 @@ describe('<Nested />', () => {
             </>
           );
         }}
-      </FormState>,
+      </FormState>
     );
-
     const input = form.find(Input);
     trigger(input, 'onChange', newAdjective);